refactor(cpp_analyzer): extract SingleFileIndexingPipeline
